add get-string-of-mesages.js (all work)
