+++ js/data.js
@@ -1,4 +1,5 @@
 import {getRandomInteger, getProgressNumber, getRandomFromArr} from './util.js';
+import {getMesseges} from './get-string-of-messages.js';
 
 const arrDescriptions = [
   'Утренний туман', 'Закат на пляже', 'Цветущее поле', 'Горный пейзаж', 'Прогулка в парке', 'Девушка на лужайке',
@@ -13,32 +14,9 @@ const arrNames = [
   'Рикардо', 'Ромина', 'Сабрина', 'Сильвия', 'Фабио', 'Франческа'
 ];
 
-const arrMessages = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
 const getIdPhoto = getProgressNumber(); // Функция возвращения случайного ID фото
 const getUrlPhoto = getProgressNumber(); // Функция возвращения случайного N URL
 const getIdComment = getProgressNumber(); // Функция возвращения случайного ID коммента
-
-/**
- *
- * @returns строку комментария
- * @description создаём из 1 или 2 случайных фраз из массива
- */
-const getMesseges = function() {
-  let strokeMesseges = '';
-  const countMessages = getRandomInteger(1, 2);
-  for(let j = 0; j < countMessages; j++) {
-    strokeMesseges += `${getRandomFromArr(arrMessages) } `;
-  }
-  return strokeMesseges.slice(0, -1);
-};
 
 /**
  *
